refactor(repeat): introduce strategy pattern
Closes #125
