+++ src/repeat.js
@@ -2,6 +2,7 @@
 import {inject} from 'aurelia-dependency-injection';
 import {ObserverLocator, calcSplices, getChangeRecords} from 'aurelia-binding';
 import {BoundViewFactory, ViewSlot, customAttribute, bindable, templateController} from 'aurelia-templating';
+import {CollectionStrategyLocator} from './collection-strategy';
 
 /**
 * Binding to iterate over an array and repeat a template
@@ -14,7 +15,7 @@ import {BoundViewFactory, ViewSlot, customAttribute, bindable, templateControlle
 */
 @customAttribute('repeat')
 @templateController
-@inject(BoundViewFactory, ViewSlot, ObserverLocator)
+@inject(BoundViewFactory, ViewSlot, ObserverLocator, CollectionStrategyLocator)
 export class Repeat {
   /**
   * List of items to bind the repeater to
@@ -26,13 +27,14 @@ export class Repeat {
   @bindable local
   @bindable key
   @bindable value
-  constructor(viewFactory, viewSlot, observerLocator) {
+  constructor(viewFactory, viewSlot, observerLocator, collectionStrategyLocator) {
     this.viewFactory = viewFactory;
     this.viewSlot = viewSlot;
     this.observerLocator = observerLocator;
     this.local = 'item';
     this.key = 'key';
     this.value = 'value';
+    this.collectionStrategyLocator = collectionStrategyLocator;
   }
 
   call(context, changes) {
@@ -40,18 +42,20 @@ export class Repeat {
   }
 
   bind(bindingContext) {
-    this.bindingContext = bindingContext;
-
-    if (!this.items) {
+    let items = this.items;
+    if (!items) {
       return;
     }
 
+    this.collectionStrategy = this.collectionStrategyLocator.getStrategy(this.items);
+    this.collectionStrategy.initialize(this, bindingContext);
     this.processItems();
   }
 
   unbind() {
+    this.collectionStrategy.dispose();
     this.items = null;
-    this.bindingContext = null;
+    this.collectionStrategy = null;
     this.viewSlot.removeAll(true);
     this.unsubscribeCollection();
   }
@@ -86,262 +90,16 @@ export class Repeat {
         this.processItemsByType();
       });
     } else {
-      this.processItemsByType();
-    }
-  }
-
-  processItemsByType() {
-    let items = this.items;
-    if (items instanceof Array) {
-      this.processArrayItems(items);
-    } else if (items instanceof Map) {
-      this.processMapEntries(items);
-    } else if ((typeof items === 'number')) {
-      this.processNumber(items);
-    } else {
-      throw new Error('Object in "repeat" must be of type Array, Map or Number');
-    }
-  }
-
-  processArrayItems(items) {
-    let viewFactory = this.viewFactory;
-    let viewSlot = this.viewSlot;
-    let i;
-    let ii;
-    let row;
-    let view;
-    let observer;
-
-    this.collectionObserver = this.observerLocator.getArrayObserver(items);
-
-    for (i = 0, ii = items.length; i < ii; ++i) {
-      row = this.createFullBindingContext(items[i], i, ii);
-      view = viewFactory.create(row);
-      viewSlot.add(view);
-    }
-
-    this.callContext = 'handleSplices';
-    this.collectionObserver.subscribe(this.callContext, this);
-  }
-
-  processMapEntries(items) {
-    let viewFactory = this.viewFactory;
-    let viewSlot = this.viewSlot;
-    let index = 0;
-    let row;
-    let view;
-    let observer;
-
-    this.collectionObserver = this.observerLocator.getMapObserver(items);
-
-    items.forEach((value, key) => {
-      row = this.createFullExecutionKvpContext(key, value, index, items.size);
-      view = viewFactory.create(row);
-      viewSlot.add(view);
-      ++index;
-    });
-
-    this.callContext = 'handleMapChangeRecords';
-    this.collectionObserver.subscribe(this.callContext, this);
-  }
-
-  processNumber(value) {
-    let viewFactory = this.viewFactory;
-    let viewSlot = this.viewSlot;
-    let childrenLength = viewSlot.children.length;
-    let i;
-    let ii;
-    let row;
-    let view;
-    let viewsToRemove;
-
-    value = Math.floor(value);
-    viewsToRemove = childrenLength - value;
-
-    if (viewsToRemove > 0) {
-      if (viewsToRemove > childrenLength) {
-        viewsToRemove = childrenLength;
-      }
-
-      for (i = 0, ii = viewsToRemove; i < ii; ++i) {
-        viewSlot.removeAt(childrenLength - (i + 1), true);
-      }
-
-      return;
-    }
-
-    for (i = childrenLength, ii = value; i < ii; ++i) {
-      row = this.createFullBindingContext(i, i, ii);
-      view = viewFactory.create(row);
-      viewSlot.add(view);
-    }
-  }
-
-  createBaseBindingContext(data) {
-    let context = {};
-    context[this.local] = data;
-    context.$parent = this.bindingContext;
-    return context;
-  }
-
-  createBaseExecutionKvpContext(key, value) {
-    let context = {};
-    context[this.key] = key;
-    context[this.value] = value;
-    context.$parent = this.bindingContext;
-    return context;
-  }
-
-  createFullBindingContext(data, index, length) {
-    let context = this.createBaseBindingContext(data);
-    return this.updateBindingContext(context, index, length);
-  }
-
-  createFullExecutionKvpContext(key, value, index, length) {
-    let context = this.createBaseExecutionKvpContext(key, value);
-    return this.updateBindingContext(context, index, length);
-  }
-
-  updateBindingContext(context, index, length) {
-    let first = (index === 0);
-    let last = (index === length - 1);
-    let even = index % 2 === 0;
-
-    context.$index = index;
-    context.$first = first;
-    context.$last = last;
-    context.$middle = !(first || last);
-    context.$odd = !even;
-    context.$even = even;
-
-    return context;
-  }
-
-  updateBindingContexts(startIndex) {
-    let children = this.viewSlot.children;
-    let length = children.length;
-
-    if (startIndex > 0) {
-      startIndex = startIndex - 1;
-    }
-
-    for (; startIndex < length; ++startIndex) {
-      this.updateBindingContext(children[startIndex].bindingContext, startIndex, length);
-    }
-  }
-
-  handleSplices(array, splices) {
-    let removeDelta = 0;
-    let viewSlot = this.viewSlot;
-    let rmPromises = [];
-
-    for (let i = 0, ii = splices.length; i < ii; ++i) {
-      let splice = splices[i];
-      let removed = splice.removed;
-
-      for (let j = 0, jj = removed.length; j < jj; ++j) {
-        let viewOrPromise = viewSlot.removeAt(splice.index + removeDelta + rmPromises.length, true);
-        if (viewOrPromise instanceof Promise) {
-          rmPromises.push(viewOrPromise);
-        }
-      }
-      removeDelta -= splice.addedCount;
-    }
-
-    if (rmPromises.length > 0) {
-      Promise.all(rmPromises).then(() => {
-        let spliceIndexLow = this.handleAddedSplices(array, splices);
-        this.updateBindingContexts(spliceIndexLow);
-      });
-    } else {
-      let spliceIndexLow = this.handleAddedSplices(array, splices);
-      this.updateBindingContexts(spliceIndexLow);
-    }
-  }
-
-  handleAddedSplices(array, splices) {
-    let spliceIndex;
-    let spliceIndexLow;
-    let arrayLength = array.length;
-    for (let i = 0, ii = splices.length; i < ii; ++i) {
-      let splice = splices[i];
-      let addIndex = spliceIndex = splice.index;
-      let end = splice.index + splice.addedCount;
-
-      if (typeof spliceIndexLow === 'undefined' || spliceIndexLow === null || spliceIndexLow > splice.index) {
-        spliceIndexLow = spliceIndex;
-      }
-
-      for (; addIndex < end; ++addIndex) {
-        let row = this.createFullBindingContext(array[addIndex], addIndex, arrayLength);
-        let view = this.viewFactory.create(row);
-        this.viewSlot.insert(addIndex, view);
+      this.collectionObserver = this.collectionStrategy.getCollectionObserver(items);
+      this.collectionStrategy.processItems(items);
+      if(this.collectionObserver) {
+        this.callContext = 'handleChanges';
+        this.collectionObserver.subscribe(this.callContext, this);
       }
     }
-
-    return spliceIndexLow;
   }
 
-  handleMapChangeRecords(map, records) {
-    let viewSlot = this.viewSlot;
-    let key;
-    let i;
-    let ii;
-    let view;
-    let children;
-    let length;
-    let row;
-    let removeIndex;
-    let record;
-
-    for (i = 0, ii = records.length; i < ii; ++i) {
-      record = records[i];
-      key = record.key;
-      switch (record.type) {
-      case 'update':
-        removeIndex = this.getViewIndexByKey(key);
-        viewSlot.removeAt(removeIndex, true);
-        row = this.createBaseExecutionKvpContext(key, map.get(key));
-        view = this.viewFactory.create(row);
-        viewSlot.insert(removeIndex, view);
-        break;
-      case 'add':
-        row = this.createBaseExecutionKvpContext(key, map.get(key));
-        view = this.viewFactory.create(row);
-        viewSlot.insert(map.size, view);
-        break;
-      case 'delete':
-        if (!record.oldValue) { return; }
-        removeIndex = this.getViewIndexByKey(key);
-        viewSlot.removeAt(removeIndex, true);
-        break;
-      case 'clear':
-        viewSlot.removeAll(true);
-        break;
-      default:
-        continue;
-      }
-    }
-
-    children = viewSlot.children;
-    length = children.length;
-
-    for (i = 0; i < length; i++) {
-      this.updateBindingContext(children[i].bindingContext, i, length);
-    }
-  }
-
-  getViewIndexByKey(key) {
-    let viewSlot = this.viewSlot;
-    let i;
-    let ii;
-    let child;
-
-    for (i = 0, ii = viewSlot.children.length; i < ii; ++i) { // TODO (martingust) better way to get index?
-      child = viewSlot.children[i];
-      if (child.bindings[0].source[this.key] === key) {
-        return i;
-      }
-    }
+  handleChanges(collection, changes) {
+    this.collectionStrategy.handleChanges(collection, changes);
   }
 }
